Files: Save the indexed information in Xapian
Save a copy of all the information as a json string in Xapian. This json
string can be used to fetch the data back when it needs to be read. We
need this because one cannot construct indexed data back into its
original form.

They normal text data is stored without any prefix while other data is
prefixed with the key's name in upper case.