File: Improve Indexing Architecture
There now exists a CommitQueue where all the xapian documents are
forwarded. When they need to be committed a xapian writable database is
created and they are committed. This had to be done so that
baloo_file_extractor did not clash with baloo_file as 2 processes cannot
have a Xapian::WritableDatabase open at the same time.