KCM: Remove email indexing option
Ideally, this should have been ported to control the new baloo pim
indexer. However, given how awesome the pim indexer is, we do not want
to ever disable it. Additionally, it is soon going to be used to provide
conversation support for emails, so it will become a hard dependency.

No point porting code which will just be thrown away