Do not call fillQueue when the file queue is created, but when the basic queue is done.
We only know what work needs to be done in the file queue once the basic
queue is complete, so call fillQueue at that time, not when the queue is
created.

Before we could be filling the file queue pointlessly with
tasks to be cleaned, or, if the file queue was empty when
the basic queue started, it might not start when it should have done.