File: Do a full check on a file in shouldIndex
For optimizations reasons we only checked if the filename should be
indexed and did not take hidden directory and location into
consideration. This was fine because files are generally only added by
the BasicIndexingQueue which does all these checks for a folder, but
when a file is sent for indexing from the FileWatch these tests are not
performed, and the file is often indexed.

From now on the tests are always performed, instead of performing them
in the FileWatcher before pushing for indexing.