Fix a potential infinite indexing loop
Minor fix to the deletion of stale documents: if they had been removed
from the FileMapping db, but remained in the Xapian db with indexing
level 1, we would previously re-index them repeatedly.

This happened because the non-existent id would be mapped to a url, eg,
$HOME/1234, which did not exist, and the extractor would exit normally.

Change it so that something which is not found anywhere is removed as if
it were an id. If the id doesn't exist in the db, nothing will be
removed.

REVIEW: 115006