BasicIQ: Remove currentUrl/Flag functions
The currentl url and flags are not used anywhere. They are just a relic
of the ancient times when file indexing was so slow that one could see
exactly which file was being indexed.

Currently, it's fast enough that the user does not really care about
which file is being indexed. Additionally that information is no longer
shown to the users.