FileIndexerConfig: Disable initial update for files
When baloo_file starts it looks through every file in your filesystem
which should be indexed and compares the mtime of each of them with the
mtime stored in the db.

This is useful to detct changes which occured when baloo_file was not
running. However, baloo should always be running so we don't really need
to check this on startup. It causes large amount of disk io on startup.