KCM: Remove the mimetype filtering and filename exclude patterns
This is going to be slightly controversial since it is removing features
which Nepomuk had, but the normal users don't care about most of these
"advanced" features.

I have kept an "Enable Developer Mode" option though. That controls if
source code should be indexed.

In the future we can create a special Baloo index tweaking application
which showcases allows advanced users to configure parts of the indexer.
The backend code still supports it, so that shouldn't be a problem.