Make sure baloo_file is just about searching
In the Nepomuk days one could enable/disable Nepomuk independent of the
file indexer. With Baloo, baloo is just about searching. The extra file
metadata which is stored is saved in the extended attributes of the file
and does not require baloo_file to be running.

This patch drops support for running only the file indexer part of
Baloo. It's now all or nothing. Additionally -

* FileWatcher - Only install watches in the indexed directories instead
of the entire home directory.

* KCM - Remove the concept of enabling/disabling Baloo. If everything is
excluded from indexing then baloo_file will not run.