File: Add a thin wrapper around different xattr implementations.
OS X, Linux and some BSDs all support extended file attributes in a way or
another, but the API differs among the operating systems.

Introduce baloo_getxattr() and baloo_setxattr(), named after the OS X and
Linux APIs for getting and setting extended attributes, to perform the right
calls depending on the operating system.

These functions accept as parameters the arguments that are common across
all xattr implementations -- this means one cannot pass custom |flags| on
Linux or |options| on OS X, but that should be fine for now as the code was
not doing that anyway.

Additionally, instead of taking raw char and void pointer, these wrappers
take QStrings are arguments, which allow us to simplify the calling code and
keep the QString->QByteArray->const char* conversions in a single place.

REVIEW:		115878