Reduce memory usage of the akonadi indexer
Most of the memory seems to be take by keeping all of the akonadi data
(mostly kmime) in memory and then indexing it. We now use the
EmitItemsIndividually delivery option and only keep one email in memory
at a time. This drastically reduces the memory usage while having
resulting in a minor decrease in speed.

A very simple test of 130k emails -
EmitItemsIndividually - 57.6 seconds - 60 mb
EmitItemsInBatches    - 56.3 seconds - 500+ mb