Extractor: Catch all exceptions
I'm not sure why an exception would be thrown over here, but we have
bugs reports which indicate that they are being thrown. One could just
let the continue crashing. The other alternative is to just abort, but
if the exception keeps getting thrown, the process will just keep
terminating and then getting respawned.