FileMap Db: Do not set the id as auto-increment
When a column is set as auto-increment an additional table
sqlite_sequence is created which stores the highest rowid. From that
point a rowid < that number is not accepted. We do not care that the
rowids be always increasing. We just need them to be unique and that is
guarenteed.

This saves us a about 16 mb when running through the
basicindexingqueuetest for 11k files.