IndexScheduler: Update everything when the config changes
During the nepomuk days doing an update was very expensive and we
therefore tried to update the minimum possible. This was hard to do and
was not perfect.

With Baloo we're fast enough to do a full scan on config changes.
Additionally the configuration does not really change very often.