MetaDataMover: Remove internal queue and make it sync
The MetaDataMover is fast enough that it does not need to be
asynchronous. It's mostly just executing a single SQL query, and even in
the cases that it does block, that's okay. All other operations will
continue after that.

Also, it was never really asynchronous it just queued the operation and
executed it later in the same thread.

REVIEW: 116919
BUG: 329039