XapianDatabase: Make the commit function blocking
The commit function now keeps trying to obtain the write lock, each time
it fails it sleep for a short interval. Calling a slot at a later point
and using a timer is not always practical.

Additionally since everyone is using the XapianDatabase class, everyone
will hold the lock for the minimum amount of time.