XapianDB: Implement a write only mode
In this writeonly mode, we do not keep 2 databases - one for reading and
one for writing. The disadvantage is that no one else can write as long
as the object is in use.

This patch also modifies the baloo_file_extractor to use this write only
db as it only need to hold the lock for a very small amount of time.