Try harder to generate up to 10 matches in Baloo's runner.
Baloo's runner plugin doesn't want to return more than 10 matches,
and was achieving this boundary by looking for up to five files or
mails and combining the results. This means at most five items were
returned if there's either no files or no mails in the index, nor
did it try to fill up that 10 if there's less than five matches of
one type.

With the approach implemented here, it'll try to fill up the full
ten by searching for as many matches as necessary, at the cost of
initially searching for up to 10 mails and potentially throwing
some of the results away (matches are still presented in files,
mail order, though). This comes on advice from Vishesh that look-
ing for mails is faster than looking for files.

Conflicts:
	src/runner/baloosearchrunner.cpp

Cherry-picked from 8df8be6ac