Extractor: Check if the file should be indexed
The check is a file should be indexed was intitally only performed
during the first phase. With this we also do it in the second phase.

This is useful as if the cleaner stops running (killed, crashed, system
shut down) it does to respawn itself and stuff continued to get indexed.
With this the extractor will clean up all the files which should not be
indexed.