ActiveFileQueueTest: Fix the test
There is a chance that when enqueing a file it gets emitted almost an
entire second earlier. This is because it could have been queued just a
couple of milliseconds before the next timer event.

It's not really a big issue.

Also, add some more debugging statements in the BasicIndexingQueueTest