Scheduler: Make sure the fileIQ is paused when on battery
The queue was getting properly paused but the basicIQ was still trying
to fill it when it finished indexing, which made so sense since the data
still had not been committed. Now each time the scheduling code is
called, the fileIQ is filled if it is supposed to be run.

Also, renamed some functions in the IndexScheduler.