FileSearch: Allow searching with wildcards in the filename
With this when searching explicitly for the property "filename", the
value can except some standard glob terms such as "*", "." and "?".
These globs can appear anywhere. This operation is NOT cheap, and it is
implemented by running a regular expression over each filename in the
database. It should only be used as a last resort.

This does not solve this bug as this is only activated when explicitly
searching for the filename. We still need to hook this up to the
frontend for the user to be able to choose "search in filename". This is
done in dolphin, but it might be nice to do it in other places such as
"krunner" as well.

CCBUG: 333037