FileModifyJob: Do not update the xapian db
This is no just a wrapper over the xattr of a file. We just set the
xattr of the file, and if baloo_file is running it will pick up the
changes and update the index. This way, there is always (almost) just
one process which is writing to the xapian db.