TermGenerator: Use QTextBoundaryFinder instead of Xapian's term generator
This gives us more control over how we split the words. However, we're
currently loosing positional information. This can easily be
implemented, but maybe it's not worth it. The database size does double
with positional information, though it does improve the quality of a
number of queries. Also, phrase queries are impossible without it.