QueryParser: Use $ instead of % in the pattern strings
We use placeholders such as '%1' in order to indicate the kind of
pattern we want to match. Eg - '%1 years'. The translators are then free
to move the '%' around if in the language the number comes after the
years.

However, using %1 in the strings causes KLocalizedString to replace the
%1 with KI18N_ARGUMENT_MISSING when not compiled in debug mode. This is
incorrect as we want the %1 to be there in the final string. With this
patch we now use $ instead of %.

REVIEW: 119519