FileWatch: Monitor attribute change events and send those for xattr indexing
This is not ideal since now when a file is touched it is sent for
indexing twice, once when the xattrs change and once when the
modification queue times out. This happens because the mtime counts as
an "attribute". The kernel makes no differnce between normal attributes
and extended attributes.