BasicIQ xattr: Do not update a doc if there are no xattr
We check to see if there were any previous xattr indexed or if there are
any new xattr and only then do we update the document. This is still not
ideal as we are still having to load the entire document on an index change
and that is not very cheap.