FileWatch: Add support for merging multiple events
Each time we get an inotify event we create a pending file object which
contains those flags. This object is passed to the queue which will
merge any incoming pending events. When the queue times out then we can
evalualte based on all the events we receive what we wish to do with
this file.

We need this because when a file is modified we get an attribute change +
modified + closedOnWrite. And we get an attribute change on any xattr
change as well. Since we do not want to index the same file twice, we
aggregate these events.

Added a bunch of unit tests for the FileWatch as well. This way we know
for sure that the correct signals are being emitted.