Query: Add Sorting Options
Baloo's search stores now support sorting in a very simple manner. They
just support 3 sorting options -

* none
* auto (default)
* property

The "property" sorting depends a lot on the backend and if it can sort based on it.
The None sorting is useful in cases such as kioslaves (Dolphin) where
the app doesn't care about the sorting and is going to show all the
results anyway and then sorting them. Ditto for KMail.

Sorting is useful in the cases such as KRunner where we only show a few
results.