Introduce an AdvancedQueryParser
This is a class which parses a string query of typical key:value pairs
and converts it into a Baloo::Term. It's supposed to be different than
the Natural Query Parser as it operates on a strict syntax and does not
have translatable strings such as "Yesterday", "Today", etc.

The implementation of this AQP is quite trivial currently. In the future
we can add support for nested expression and additional operators. The
idea is that this parser can be used in the kioslave so that more
technical users can write their own queries, in QML and as a basis for
representing already constructed Baloo::Queries into user visible text.