get rid of FileIndexingJob, use ExtractorClient instead
drastically simplify and limit the collection class handling in
FileIndexingQueue by moving to ExtractorClient. also starts using
the sql database to track indexing state of urls, so no more
closing/opening the xapian db over and over again or writing to it
either.

i suspect that the sqlite usage is not optimal either as it gets hit
once per file indexed, but that needs to wait on profiling to
determine.