Escape dots in exclude filters
In regular expressions a dot (.) matches any character (except newline)
but in the exclude filters we use wildcard syntax (*) and want a dot (.)
to be interpreted as a character.

Example: With "*.o" in the exclude filters the user expects object
files (ending with .o) are excluded. Without escaping this would match
every file and folder ending with o though. This is the case for all
entries of that form in exlude filters ("*.moc", "*.la", etc.)

So just escape every dot we find in exclude filters with a backslash
while building the regexp.

BUG: 339908
FIXED-IN: 5.1.1
REVIEW: 120570