Add "type" and "kind" properties
The type is generally exposed via the Query class and a dedicated
setter/getter. However, it's useful to be able to specify it in the term
as well. Also, this way the query parser will recognize them as
properties and allow users to create search folders based with "type" or
"kind".

type and kind are the same thing.