Introduce a FilteredDirIterator class
This FilteredDirIterator iterates over a folder recursively ignoring all
the files/folders which should not be indexed. Both kinotify and the
BasicIndexingQueue do this on their own. Also, each of them does not
maintain an internal state and runs all the regular expresions on the
entire path each time. This is incredibly inefficient as we check large
parts of the path multiple times.

This FilteredDirIterator currently also checks the full path each time.
It can be improved in another patch.