KInotify: Use FilteredDirIterator
The FilteredDirIterator is a lot more efficient at iterating over the
entire file sytem which ignoring folders which should not be indexed.
This results in KInotify being less generic and more tied to Baloo, but
it is probably worth it since we will be reducing code duplication and
getting a huge performance boost during startup.

This patch removes the IgnoringKInotify class which was subclassed from
KInotify. It's not required since KInotify now has the config.