kinotify: Remove internal buffer
We kept an internal buffer to which we read the data into. This buffer
was of a very large size (about 16k) and was always kept of the stack.
Considering that we can request the size of the buffer, we do not really
need to have such a large size.

Also, this might remove some of the critical warnings we were getting
with the coverity scan of potential illegal memory access, and
insecure data access.