Stop using sqlite
We now instead store the full path in Xapian. It is once stored in a
slot value so that it can be efficiently retrieved, and once it is
stored as a term so that we can use that to lookup the id.

This results in the metadata mover's code becoming quite complex as it
can now no longer rely on the sql engine to update the paths when a
folder is moved. Instead we now need to manually iterate over all the
files and do it ourselves.

The PathFilterPostingSource is now also a lot slower as it is checking
each term. We can possibly optimize it more.

Overall, this does simplify the code in certain places as we no longer
need to bother about maintaining 2 databases and keeping them in sync.