Store the sha1hash of the url in xapian
We store the entire URL in the values of Xapian, and we store the url as
a term. Storing it as a term allows us to do a lookup for url -> id.
However, xapian can only store terms of up to 245 characters. Quite
often the entire path is much larger than 245 characters. Therefore we
cannot store it there.

Instead we store the cryptographic hash of the url. This way the lookup
still works, except that now we cannot iterate over all the urls which
begin with a certain path, and we need that for the metadata mover.

One test fails with this patch.