Add a positiondb
This can be used to store terms along with both their docId and the
positions which they appear in. I've chosen to duplicate the entire
inverted index once for the posting lists (term -> [docID]) and once for
positions (term -> [docID + positions]). One can argue that with this
positiondb one doesn't really need the postingdb. And they would be
correct. However, most queries would not need the additional overhead of
fetching the positions.