Add an IndexingLevelDB
In Baloo, we have 2 phase indexing. During the first phase only the file
name + mtime + mimetype are indexed. During the second phase the file's
contents are indexed.

We need some way of marking files which have completed the first phase
of indexing so that they can be sent for the second phase. With Xapian,
we did this using a term "Z1". This would result in the posting list of
"Z1", to be huge. And each time a document would have gone through
basic indexing, the posting list containing every document would need to
be modified. Xapian has internal optimizations for handling very large
posting lists, so this hasn't been a huge problem, but it is still a
sore point.

Now that we are writing our own engine. We can have a separate btree for
storing the document ids which have completed basic indexing. It'll be
far more efficient this way.