Integrate the DocUrlDb and UrlDocDb with the Database
You can now set the url in the Document class and that will then be
updated accordingly in the appropriate databases.

Also added members for mapping docId <-> docUrl in the database class.
We need to do this very frequently, and it makes sense to not have to
fetch the entire document when that needs to be done.