Port CommitQueue
The way I have implemneted this is a bit hackish. If a Doc id is provided to it, the url of the doc is read and that is used to replace document in the index, else
doc is simply added. Now since doc id is useless we can maybe replace it with a bool simply, which tells us to add a new doc or replace existing.