XapianDatabase: Remove previous data before adding
When moving files across different file systems we do not get move
events as we are actually deleteing + creating a new file. In this case
we get a xapian delete request and then when it gets indexed an add
request. Since we batch together requests, this resulted in the new data
being added and then delete, if the id was same. The id sometimes is
same if no other file is being added.

Ideally, we should be grouping the delete + adds acording to timestamp,
but that is for another patch.

This seems to fix moving file across different mount points for me.

BUG: 33422
FIXED-IN: 5.2.1