Set the file id to be the (dev id + inode number)
Instead of using a monotonically increasing integer, we're using the
inode number and device id. On a 64-bit system - both these numbers seem
to be 64 bit. However, given we do not need such large identifiers,
we're converting them to 32-bit integers.

This will allow us to get rid of the (file url -> id) mapping database,
as we can just rely on the underlying filesystem. Additionally, this
makes Baloo a more useful tool in the global linux world as we can map
inode numbers to file paths.