BasicIQ: Avoid using the FileMapping
The FileMapping class is not that useful anymore. It used to provide a
convenient wrapper on top of (file path <-> file id), but now that we
can rely on the file system to help us convert the file path -> file id,
the basic IQ becomes much simpler, and faster.

This increases the speed about the basic IQ by about 10-15%.