Make the DocumentUrlDB not store the entire url
Instead it uses the IdTreeDb and IDFilenameDb to only store the filename
along with the parent folder's doc id.

We need more tests for this since it isn't well tested, and the code is
decently complex with many edge cases. Plus, since the DocumentUrlDB now
stat's every part of the url, it's quite expensive and introduces a hard
dependency with the filesystem, which makes testing much harder.