DocumentUrlDB: Avoid storing parts of the path multiple times
If storing a url "/home/vish/A", we need to create entries for "A",
"vish" and "home". Next, when storing "/home/vish/B", we do not need to
recreate entries for "home" and "vish".

This improves the speed of inserting urls substantially (43 sec to 13
sec for 700k files). We are still doing a few extra stat calls, but
those can be optimized later.