BasicIndexingJob: Do not store the time as terms
We have a mtimedb which can be used to do lookups. We do not need to
store every part of the mtime seperately. Though we do loose the ability
to only lookup the day / month without the year.

This reduces indexing time significantly and reduces the database size
by a good 30%.