Remove the Term class from the public
It's better to expose the minimum number of classes as a public API. The
Query class was quite complicated as it offerred both Terms and a search
string along with types. Since all the functionality of the terms can be
achieved with the search string, it's better to remove the Term class
for now.

Later we can reintroduce it in another manner if required.

This breaks API.