Engine: Make individual database classes short lived
Each Database class is now more of a namespace with a given db id and
transaction. So the database class just stores all the db ids, but not
all the smallers databse classes.

Each time it needs to perform any operation it creates those database
classes on the stack, and give them the db num and transaction.

The main reason for doing this is so that we can make the database
multi-threaded. We need multiple threads to perform independent
transactions.