Remove the CommitQueue
The CommitQueue was used to batch up operations and then commit them
after an interval. While the idea was great, now with the new engine,
our transactions need to be short lived. We cannot just throw everything
into one big queue.

The BasicIQ is now extremely slow as it is performing a commit after
each file. This needs to be fixed.