Engine: Expose the toPostingIterator function
We need to expose the following additional features -
* mtime posting iterators
* id tree posting iterators

Instead of extending EngineQuery to support all of this and any other
possible extensions, it makes more sense for clients to directly
interact with the PostingIterator and compose the query they want.

We want to avoid adding another abstraction. We already have the
following the for an end user -

Text -> Advanced Query Parser -> Baloo::Term -> Search Store -> Engine
Query -> Database -> PostingIterators -> Raw Data.

It makes sense to allow the SearchStore to directly interact with the
PostingIterators when they need it.