Add an UnIndexedFileIterator
This provides an iterator which can be used to only iterate over all the
files under a certain path which need to be reindexed. This internally
used the FilteredDirIterator which is quite efficient since it doesn't
check the entire file path each time and only checks the changes.

We can use this UnIndexedFileIterator in the BasicIQ later and
substantially improve performance.