Introduce Database open modes - open or create
This way read only Database never create a write transaction since they
always only open the database, but never create it.

Before this when a transaction was running, you could not run a query
with `baloosearch` since that would involve also creating a write
transaction and try to create the databases.