Extractor: Handle id changes gracefully
There can be a case when baloo_file is not running or is stuck, and a
file is deleted and a new one with the same name is created. In this
case the file id will have changed since it will now have a different
inode number.

In this case we should update our db, and remove the document with the
id number that no longer exists.