Do not expose the file id
We can keep the file identifier as an internal id which is not exposed.
All the user actually cares about is the file path. Also, this way we
obtain all the filepaths in the same transaction instead of starting a
different transaction to get each file path. Specially since when we are
requesting the file paths, the files could be deleted/modified.