Revert "Only use the file's content during mimetype detection"
This reverts commit c19b7a9ded994009c49007d8336afe92acf513cd.

By only using the file's contents we do not get the proper mimetype for
certain text files such as 'CMakeLists.txt'. The problem with this is
that we internally have a hack to not index files with mimetype
'text/plain' unless they with a '.txt'.

Therefore this will break indexing of certain files in many edge cases.
Best not to push this in a bug fix release. We can investigate all the
effects of this in master, since the next release is in August.