Handle _ as the special case that it is
Words are not split based on _ via QTextBoundaryFinder which uses
Unicode Standard Annex #14 and #29. However, that's kinda idiotic.
We need words to be split by '_'.

If a filename contains 'plant_name' searching for 'name' should give us
that file. We handled this case with a special word.split('_') at the
end. This however did not target the case when the word starts with an _
as then QTextBoundaryFinder does not consider it an item and completely
ignores it.

We now therefore just always convert all '_'s into ' 's