BasicIndexinJob: Consider the mimetype as text instead of a blob
When searching for the mimetype, the queryparser splits the query into
multiple smaller queries, as it should. Example - 'video/mp4' -> (video,
1) + (mp4, 2). If we save the mimetype as a blob, then these queries
will not work as the text is not split.

We could hack around this by having a special case for the mimetype, but
it doesn't seem like it is worth it. Though this would slow down the
initial indexing a little bit, and increase the size of the db.