Disable compiling the Baloo Daemon without inotify
Baloo requires inotify. It's a hard dependency for us. When we do not
have it, we connect to kio signals. But that code path isn't tested
and only provides a sub-set of the required features. Baloo might be
somewhat usable on those systems, but we really do not know. The idea of
shipping a half baked product does not seem appealing. It seems better
to focus on what we can do, and do that well.

Baloo will still compile its library on non-Linux based systems, since
its a dependency for some applications (eg - Dolphin), but it will not do
anything.

CCMAIL: release-team@kde.org