MetdataMover: Use the DocumentUrl to map path -> id instead of the fs
The fs has already changed at this point, it's not wise to use it as
another file/folder with the same name could have been created in some
part of the part, and therefore when trying to resolve the path -> id we
may get an incorrect id. It's better to just use the db, even though it
would be much slower.

This also exposes how slow removing a document from the database is. It
takes a good 4-500 msecs per document. This kind of speed is obviously
not acceptable. It was never noticed before because all the documents
were not being removed.