MetaDataMover: Recursively remove directories
Even though we are supposed to get events for the all the file deletions
before a folder is deleted, it's wiser for us to do recursive deletes
incase we missed a notification. Also, in the future, perhaps we want to
avoid sending notifications for all the file changes if the parent
folder is just going to be deleted.