FileIndexerConfig: Do not update the config automatically on fs change
The File System and its events are not reliable for us. It's not that we
do not get the signal, we often get too many signals, which cause us to
restart indexing even though something might not have changed.

It's easier for the KCM to just explicitly tell us when it has updated
the config.