Add a test for RemoveRecursively
Also remove an invalid assert. There can be cases when a document is
being deleted which has no terms. This happens cause we mirror the file
system structure, but do not index all of it.

Eg - /home/vish/unindexedFolder/a/indexedFolder/fire

/home/vish is indexed
/home/vish/unindexedFolder is NOT indexed
but the file 'fire' is indexed.

When 'fire' is removed, we will also be clearing the records for
'unindexedFolder' which does not have any terms since it was not
indexed.

Yes, this is confusing.