Hook up fileRemoved signals from the FileWatcher to the Scheduler
The FileWatcher can notify us when it is removing files, this way we can
update the queues in the scheduler so that files which it has removed
are no longer in the queue to be indexed. This is especially useful when
a large directory is created and then later removed before it can be
indexed.