Revert to previous behaviour of getting URL from Extractor
Querying a property can be delayed by any amount of time therefore it can
screw up synchronization and lead to one URL to be fetched more than once
or not fetching a name at all as it has changed to something else by the time
the fetch call is returned by the extractor. We're still keeping the property
around as it is useful to fetch the current file when a monitor starts.