SearchStore: Remove asserts for document url checks
Our database is not always in a consistent state. Somehow we remove URLs
from the documenturldb but we do not remove its terms. I have no idea
why this happens, and I don't understand it.

The assert was so that I catch this at run time and hopefully fix it.
But since Baloo is now being used more in production, we cannot afford
Plasma and KRunner to crash with these asserts.

This is not good.

BUG: 350246