ContentIndexer: Handle a strange case better
The IOHanlder in the extractor has a check for if 0 number of ids are
passed. Therefore, I've added a similar assert in
ExtractorProcess::index.

The only time when this can happen is if the provider tells us there are
files to index, and on requesting them doesn't give us the files because
the database has changed in the middle.