Schduler: Remove debug output
It seems on certain distros the debug output goes in .xsession-erros,
for which we then get file modified notifications and it is sent for
indexing and then ignored. Debug output is generated on the way, which
is then triggers this whole cycle again.

Disabling debug output is probably not the best solution, as this output
is useful.