Prevent baloo_file_extractor being killed while committing
Since we emit the filenames before we start indexing them, for diagnostic reasons,
simply counting the number of files written to stdout is not enough we need to make
sure baloo_file_extractor has commited before it is destroyed (when indexing the last
batch or suspending it). So now we wait for baloo_file_extractor to write "batch commited"
to stdout before the ExtractorProcess emits done.

REVIEW: 124836