Remove KAuth support to increase max inotify watches if count too low
If the maximum number of inotify watches is too low, we would show a
KAuth dialog asking the user for the root password so that we could
increase the number of watches. Lots of users seem to hate this since
they feel uncomfortable for Baloo to be asking for the root password.

Given that this is an enviroment problem, I'm disabling this extra
support for increasing the limit. Baloo will ignore changes on files
when this happens. In the future, I'll probably just disable Baloo
completely when this happens.

BUG: 351602