Add a FileInfo class
This FileInfo class is meant to be a replacement of QFileInfo. Our
replacement just directly calls 'stat' internally. It takes a QByteArray
instead of QString and therefore we can avoid the utf16 -> utf8
conversion.

It should also be a little bit faster as we are avoiding all the extra
logic inside QFileInfo which we do not care about.