Fix limit/offset handling in SearchStore::exec
Before offset is introduced, the result is returned within [0, limit),
with the support of offset, the result returned should be within
[offfset, offset + limit).
1. When we need to sort the result, whole result should always be sorted,
instead of being sorted partially. This is consistent with SQL ORDER BY
and LIMIT.
2. When we don't need sort, we skip number of "offset" result and read
until it reaches (offset + limit), also need to be care when limit is
less than zero (means grab all result).

REVIEW: 125544